Move logic from SnippetArea to redux actions
