+++ src/actions/app.js
@@ -2,8 +2,8 @@ import axios from 'axios';
 
 
 import { addNotification, closeNotification } from './notifications';
-import { setAuthor } from './permissions';
-import { updateUserSnippets, fetchSnippetLists } from './user';
+import { setAuthor, setPermissions } from './permissions';
+import { fetchSnippetLists } from './user';
 import { removeDeprecatedFiltersFromState } from '../util/codemirror-utils';
 import { makeSaveEndpointUrl, normalizeState } from '../util/requests';
 import { setDefaults } from '../util/state-management';
@@ -126,8 +126,14 @@ export const saveNew = org => (dispatch, getState) => {
         // Clear the hasUnsavedChanges flag
         dispatch(saveSucceeded());
         // Update the user's personal and organization snippet lists
-        dispatch(updateUserSnippets());
         dispatch(fetchSnippetLists());
+        // Update the snippet's key
+        dispatch(setSnippetKey(res.data.key));
+        const permissions = {
+          canRead: true,
+          canEdit: true,
+        }; // Grant all permissions, this is now her file.
+        dispatch(setPermissions(permissions));
         // Return the key of the newly-saved snippet so that the browser
         // location can be updated
         return res.data.key;
@@ -171,7 +177,6 @@ export const saveExisting = () => (dispatch, getState) => {
         // Clear the hasUnsavedChanges flag
         dispatch(saveSucceeded());
         // Update the user's personal and organization snippet lists
-        dispatch(updateUserSnippets());
         dispatch(fetchSnippetLists());
       }, () => {
         // Give user feedback that saving failed
@@ -203,6 +208,11 @@ export const deleteSnippet = snippetKey => (dispatch, getState) => {
       dispatch(closeNotification());
       // Give user feedback that snippet deleted
       dispatch(addNotification('Snippet Deleted!'));
+
+      // Update state to reflect deletion of snippet
+      dispatch(resetState());
+      dispatch(setAuthor(''));
+      dispatch(fetchSnippetLists()); // Update org snippet lists
     })
     .catch((err) => {
       // Remove the 'deleting...' notifications
